+++ src/Config.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sensorario\Yaphlo;
+
+interface Config
+{
+    public function level();
+}