+++ src/Message.php
@@ -12,6 +12,8 @@ class Message
 
     const LEVEL_FATAL = 'FATAL';
 
+    private ?string $channel = null;
+
     private ?string $level = null;
 
     private array $content = [];
@@ -65,11 +67,16 @@ class Message
 
         $datetime = $this->datetime->format('[Y-m-d H:i:s]');
         foreach (explode("\n", $encoded) as $line) {
-            if ($this->level === null) {
-                $rendered[] = $datetime . ' ' . $line;
-            } else {
-                $rendered[] = $datetime . ' ' . '[' . $this->level . ']' . ' ' . $line;
-            }
+            $items = [];
+
+            $items[] = $datetime;
+
+            if ($this->level !== null) { $items[] = '[' . $this->level . ']'; }
+            if ($this->channel !== null) { $items[] = '[' . $this->channel . ']'; }
+
+            $items[] = $line;
+
+            $rendered[] = join(' ', $items);
         }
 
         return implode("\n", $rendered);
@@ -87,12 +94,17 @@ class Message
         );
     }
 
-    public function doesLevelPrints(string $level): bool
+    public function isPrintableWithLevel(string $level): bool
     {
         $map = $this->inverseMap();
         $levelNumber = $map[$level];
         $levelMe = $map[$this->level] ?? 0;
 
         return $levelNumber >= $levelMe;
+    }
+
+    public function setChannel(string $channel): void
+    {
+        $this->channel = $channel;
     }
 }