+++ tests/WriterTest.php
@@ -3,6 +3,7 @@
 namespace Sensorario\Yaphlo\Tests;
 
 use Sensorario\Yaphlo\Writer;
+use Sensorario\Yaphlo\Message;
 
 class WriterTest extends \PHPUnit\Framework\TestCase
 {
@@ -12,6 +13,11 @@ class WriterTest extends \PHPUnit\Framework\TestCase
 
     public function setUp(): void
     {
+        $this->conf = $this
+            ->getMockBuilder(\Sensorario\Yaphlo\Config::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
         $this->message = $this
             ->getMockBuilder(\Sensorario\Yaphlo\Message::class)
             ->disableOriginalConstructor()
@@ -34,7 +40,50 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             ->expects($this->once())
             ->method('render');
 
-        $writer = new Writer($this->filePutContent);
+        $this->conf
+            ->expects($this->once())
+            ->method('level')
+            ->willReturn(Message::LEVEL_INFO);
+
+        $this->message
+            ->expects($this->once())
+            ->method('isPrintableWithLevel')
+            ->willReturn(true);
+
+        $writer = new Writer(
+            $this->conf,
+            $this->filePutContent,
+        );
+
+        $writer->write($this->message);
+    }
+
+    /** @test */
+    public function doNotWriteIntoFile()
+    {
+        $this->conf
+            ->expects($this->once())
+            ->method('level')
+            ->willReturn(Message::LEVEL_FATAL);
+
+        $this->filePutContent
+            ->expects($this->never())
+            ->method('append');
+
+        $this->message
+            ->expects($this->never())
+            ->method('render');
+
+        $this->message
+            ->expects($this->once())
+            ->method('isPrintableWithLevel')
+            ->willReturn(false);
+
+        $writer = new Writer(
+            $this->conf,
+            $this->filePutContent,
+        );
+
         $writer->write($this->message);
     }
 }