+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v1.0
 
  - complete level list with INFO, WARNING, ERROR, FATAL
+ - add channel to cathegorize messages